refactor: Added a new context for extensionStates
