+++ app/scripts/preload.js
@@ -1,4 +1,11 @@
+const { ipcRenderer } = require('electron');
+
 require('./mouseTracking');
 require('./radialUI');
 require('./scrollHelp.js');
 require('./AlertMessageBookmark');
+
+window.addEventListener('DOMContentLoaded', () => {
+	ipcRenderer.send('getExtensionStates');
+});
+
