+++ src/components/Settings/components/Pages/ExtensionSetting.jsx
@@ -2,52 +2,21 @@ import React, { useEffect } from 'react';
 import Title from '../../../Typography/Title';
 import CustomSwitch from '../CustomSwitch';
 import SettingTile from '../SettingTile';
-import { doc, getDoc, setDoc } from 'firebase/firestore';
+import { doc, setDoc } from 'firebase/firestore';
 import { db } from '../../../../utils/FirebaseConfig';
 
 import { useAuth } from '../../../../contexts/AuthContextProvider';
 
 import { MinusIcon } from '@heroicons/react/outline';
 import { PlusIcon } from '@heroicons/react/outline';
-import useLocalStorageState from '../../../../hooks/useLocalStorageState';
+import { useExtensionStates } from '../../../../contexts/ExtensionStatesContextProvider';
 
 const ExtensionSetting = () => {
 	const { user } = useAuth();
-	const [extensionStates, setExtensionStates] = useLocalStorageState('extensionStates', {
-		scrollSpeed: 3,
-		radialUI: true,
-		mousePrediction: true,
-		scrollHelp: true,
-		scrollHelpPosition: { top: '68%', left: '90%' },
-	});
-
-	window.api.recieve('getExtensionStatesReply', () => {
-		window.api.send('extensionStates', extensionStates);
-	});
-
-	window.api.recieve('setLatestOverlayLocationReply', (...payload) => {
-		setExtensionStates((prevExtensionStates) => ({
-			...prevExtensionStates,
-			scrollHelpPosition: { top: payload[0][0], left: payload[0][1] },
-		}));
-	});
-
+	const { extensionStates, setExtensionStates } = useExtensionStates();
 	useEffect(() => {
 		if (!user) return;
-
-		const fetchData = async () => {
-			const docRef = doc(db, 'users', user.uid);
-			const docSnap = await getDoc(docRef);
-
-			if (docSnap.exists()) {
-				const data = docSnap.data();
-				if (data.extensionStates) {
-					setExtensionStates(data.extensionStates);
-				}
-			}
-		};
-		fetchData();
-	}, [user, setExtensionStates]);
+	}, [user, extensionStates]);
 
 	const incrementScrollSpeed = async () => {
 		if (extensionStates.scrollSpeed >= 10) return;
