+++ src/index.jsx
@@ -5,12 +5,15 @@ import { HashRouter } from 'react-router-dom';
 import AuthContextProvider from './contexts/AuthContextProvider';
 
 import './css/index.css';
+import ExtensionStatesContextProvider from './contexts/ExtensionStatesContextProvider';
 
 ReactDOM.render(
 	<React.StrictMode>
 		<HashRouter>
 			<AuthContextProvider>
-				<App />
+				<ExtensionStatesContextProvider>
+					<App />
+				</ExtensionStatesContextProvider>
 			</AuthContextProvider>
 		</HashRouter>
 	</React.StrictMode>,
